requirements.txt
Django==4.1.3
django-allauth==0.51.0
djangorestframework==3.14.0
gevent==22.10.2
gunicorn==20.1.0
